Be resilient to missing drive letters or root subdirectories.
Since Wasmtime 20, it is now an error to preopen a directory that is
not accessible by Wasmtime. In Wasmtime 20, the error would happen in
the store initialization code, which made it hard to work around, but
since Wasmtime 21, this happens at the preopen_dir() call. This was
changed in response to bytecodealliance/wasmtime#8552.

Preopens for relative paths are unaffected because if one of parents of
your cwd can't be opened by you you should probably do something about
it. This may be reconsidered in the future (perhaps with a warning) if
there are enough broken systems configured in this way to outweigh
the risk of ending up with certain paths that you can open being not
possible to use with YoWASP tools. (E.g. a -r+x directory is going to
be rejected by Wasmtime, but you could ordinarily use a path that
traverses it.)